strip operator methods from Design; move helpers into operator.rs
Design is now pure data + layout: store, factors, n_rows, n_dofs,
from_store, n_factors. The matvec helpers (gather_add, scatter_add,
scatter_add_single_factor, level_from_column_or_store, ScatterStrategy,
PAR_THRESHOLD, SCATTER_LOCAL_THRESHOLD, factor_columns) move into
operator.rs as module-private free functions taking &Design. The
DesignOperator::apply / apply_adjoint methods now call them directly.

- Removed Design::matvec_d, rmatvec_dt, rmatvec_wdt, uid_weight (and the
  internal gather_add/scatter_add methods). The uid_weight read in
  cross_tab.rs inlines to weights.map_or(1.0, |w| w[uid]).
- Solver::solve diagnostic computes D^T W v via apply_adjoint(W^{1/2} v)
  instead of the dedicated rmatvec_wdt method; results are bit-identical
  because D^T W = D^T W^{1/2} W^{1/2} algebraically.
- Tests/examples/benches that called design.matvec_d(...) etc. now use
  DesignOperator::new(&design, None).apply(...) (and likewise for the
  adjoint). The internal tests in src/domain.rs and src/operator/tests.rs
  are pruned to remove redundant matvec coverage that DesignOperator
  already exercises.
