+++ crates/within/examples/solve_demo.rs
@@ -19,7 +19,9 @@ fn main() {
     }
 
     // Build design to compute D * x_true.
+    use schwarz_precond::Operator as _;
     use within::observation::FactorMajorStore;
+    use within::operator::DesignOperator;
     use within::Design;
 
     let factor_levels = vec![categories.column(0).to_vec(), categories.column(1).to_vec()];
@@ -30,7 +32,9 @@ fn main() {
     let total_dofs = design.n_dofs;
     let x_true: Vec<f64> = (0..total_dofs).map(|j| (j % 7) as f64 - 3.0).collect();
     let mut y = vec![0.0; n_obs];
-    design.matvec_d(&x_true, &mut y);
+    DesignOperator::new(&design, None)
+        .apply(&x_true, &mut y)
+        .expect("apply succeeds");
     // Add small deterministic perturbation so the system is not trivially exact.
     for (i, yi) in y.iter_mut().enumerate() {
         *yi += 0.01 * ((i * 7 + 3) % 13) as f64 - 0.06;
